+++ components/proteins/AddProtein.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { mutate} from 'swr'
 import useForm from 'hooks/useForm';
-import { headers } from 'hooks/swr-switch';
+import { headers } from 'lib/utils';
 
 import Form from '../reusable/Form';
