+++ components/sauces/SaucePairing.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { useSauces, useWines } from 'hooks/swr-hooks';
 
-import { headers } from 'hooks/swr-switch';
+import { headers } from 'lib/utils';
 import useForm from 'hooks/useForm';
 
 import Form from '../reusable/Form';
