+++ components/userFlow/SignOut.tsx
@@ -1,14 +1,15 @@
 import { useState } from 'react';
 import {mutate} from 'swr';
 
-import { headers } from 'hooks/swr-switch';
+import { headers } from 'lib/utils';
 import useForm from 'hooks/useForm';
 
 import DeleteButton from '../styles/DeleteButton';
+import { FetchOptions } from 'types';
 
 export default function SignOut() {
-  const [isSubmitting, setIsSubmitting] = useState();
-  const [errorMessage, setErrorMessage] = useState();
+  const [isSubmitting, setIsSubmitting] = useState(false);
+  const [errorMessage, setErrorMessage] = useState('');
   const { handleSubmit } = useForm(callback);
 
   async function callback() {
@@ -16,12 +17,12 @@ export default function SignOut() {
       setIsSubmitting(true)
       try {
         const url = 'api/logout'
-        const options = {
+        const options: FetchOptions = {
           method: 'POST',
           credentials: 'include',
           headers,
         }
-        const res = await fetch (url,options);
+        const res = await fetch (url, options);
         const data = await res.json();
         if (!res.ok) {
           throw new Error(data.message)
