+++ components/wines/EditWine.tsx
@@ -3,9 +3,15 @@ import { useWine } from 'hooks/swr-hooks';
 import EditWineForm from './EditWineForm';
 import Loader from '../reusable/Loader';
 
-export default function EditWine({ id }) {
+interface Props {
+  id: number;
+}
+
+const EditWine: React.FC<Props> = ({ id }) => {
   const { data, isLoading } = useWine(id);
 
   if (isLoading) return <Loader />;
   return <EditWineForm data={data} id={id} />;
 }
+
+export default EditWine;
